Move address actions from actions.ts
